backend/model/cropDisease/requirements.txt
fastapi
uvicorn
torch
torchvision
timm
pillow
numpy
python-multipart
pandas
